[Fix] review, add demo
