+++ src/main/java/com/alipay/oceanbase/rpc/mutation/Delete.java
@@ -20,6 +20,7 @@ package com.alipay.oceanbase.rpc.mutation;
 import com.alipay.oceanbase.rpc.ObTableClient;
 import com.alipay.oceanbase.rpc.exception.ObTableException;
 import com.alipay.oceanbase.rpc.mutation.result.MutationResult;
+import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.ObTableOperation;
 import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.ObTableOperationType;
 import com.alipay.oceanbase.rpc.table.api.Table;
 
@@ -65,8 +66,10 @@ public class Delete extends Mutation<Delete> {
                 getTableName(), getRowKey(), getKeyRanges()));
         } else {
             // QueryAndDelete
+            ObTableOperation operation = ObTableOperation.getInstance(ObTableOperationType.DEL, new Object[] {},
+                    null, null);
             return new MutationResult(((ObTableClient) getClient()).mutationWithFilter(getQuery(),
-                getRowKey(), getKeyRanges(), ObTableOperationType.DEL, null, null, false));
+                    getRowKey(), getKeyRanges(), operation, false));
         }
     }
 }
