+++ src/main/java/com/alipay/oceanbase/rpc/mutation/Increment.java
@@ -20,6 +20,7 @@ package com.alipay.oceanbase.rpc.mutation;
 import com.alipay.oceanbase.rpc.ObTableClient;
 import com.alipay.oceanbase.rpc.exception.ObTableException;
 import com.alipay.oceanbase.rpc.mutation.result.MutationResult;
+import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.ObTableOperation;
 import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.ObTableOperationType;
 import com.alipay.oceanbase.rpc.table.api.Table;
 
@@ -137,9 +138,10 @@ public class Increment extends Mutation<Increment> {
                 values.toArray(), withResult));
         } else {
             // QueryAndIncrement
+            ObTableOperation operation = ObTableOperation.getInstance(ObTableOperationType.INCREMENT, new Object[] {},
+                    columns.toArray(new String[0]), values.toArray());
             return new MutationResult(((ObTableClient) getClient()).mutationWithFilter(getQuery(),
-                getRowKey(), getKeyRanges(), ObTableOperationType.INCREMENT,
-                columns.toArray(new String[0]), values.toArray(), withResult));
+                    getRowKey(), getKeyRanges(), operation, withResult));
         }
     }
 }
