+++ src/main/java/com/alipay/oceanbase/rpc/mutation/Insert.java
@@ -19,7 +19,9 @@ package com.alipay.oceanbase.rpc.mutation;
 
 import com.alipay.oceanbase.rpc.ObTableClient;
 import com.alipay.oceanbase.rpc.exception.ObTableException;
+import com.alipay.oceanbase.rpc.filter.ObTableFilter;
 import com.alipay.oceanbase.rpc.mutation.result.MutationResult;
+import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.ObTableOperation;
 import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.ObTableOperationType;
 import com.alipay.oceanbase.rpc.table.api.Table;
 
@@ -39,6 +41,29 @@ public class Insert extends Mutation<Insert> {
         columns = new ArrayList<String>();
         values = new ArrayList<Object>();
         setInsert();
+    }
+
+    /*
+     * set the Row Key of mutation with Row and keep scan range
+     */
+    @Override
+    public Insert setRowKey(Row rowKey) {
+        return setRowKeyOnly(rowKey);
+    }
+
+    /*
+     * set the Row Key of mutation with ColumnValues and keep scan range
+     */
+    @Override
+    public Insert setRowKey(ColumnValue... rowKey) {
+        return setRowKeyOnly(rowKey);
+    }
+
+    /*
+     * add filter into mutation (use QueryAndMutate) and scan range
+     */
+    public Insert setFilter(ObTableFilter filter) throws Exception {
+        return setFilterOnly(filter);
     }
 
     /*
@@ -135,9 +160,11 @@ public class Insert extends Mutation<Insert> {
                 getTableName(), getRowKey(), getKeyRanges(), columns.toArray(new String[0]),
                 values.toArray()));
         } else {
+            // QueryAndInsert
+            ObTableOperation operation = ObTableOperation.getInstance(ObTableOperationType.INSERT, getRowKey(),
+                    columns.toArray(new String[0]), values.toArray());
             return new MutationResult(((ObTableClient) getClient()).mutationWithFilter(getQuery(),
-                    getRowKey(), getKeyRanges(), ObTableOperationType.INSERT,
-                    columns.toArray(new String[0]), values.toArray(), true));
+                    getRowKey(), getKeyRanges(), operation, true));
         }
     }
 }
