+++ src/test/java/com/alipay/oceanbase/rpc/ObTableClientAutoIncTest.java
@@ -1,0 +1,497 @@
+package com.alipay.oceanbase.rpc;
+
+import com.alipay.oceanbase.rpc.bolt.ObTableClientTestBase;
+import com.alipay.oceanbase.rpc.exception.ObTableException;
+import com.alipay.oceanbase.rpc.filter.ObCompareOp;
+import com.alipay.oceanbase.rpc.filter.ObTableValueFilter;
+import com.alipay.oceanbase.rpc.location.model.ObServerAddr;
+import com.alipay.oceanbase.rpc.location.model.ServerRoster;
+import com.alipay.oceanbase.rpc.mutation.result.MutationResult;
+import com.alipay.oceanbase.rpc.property.Property;
+import com.alipay.oceanbase.rpc.protocol.payload.ResultCodes;
+import com.alipay.oceanbase.rpc.stream.QueryResultSet;
+import com.alipay.oceanbase.rpc.table.api.TableQuery;
+import com.alipay.oceanbase.rpc.util.ObTableClientTestUtil;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+import java.lang.reflect.Field;
+import java.sql.Connection;
+import java.sql.Statement;
+import java.util.Map;
+
+import static com.alipay.oceanbase.rpc.filter.ObTableFilterFactory.compareVal;
+import static com.alipay.oceanbase.rpc.mutation.MutationFactory.colVal;
+import static com.alipay.oceanbase.rpc.mutation.MutationFactory.row;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+
+public class ObTableClientAutoIncTest extends ObTableClientTestBase {
+    @Before
+    public void setup() throws Exception {
+        System.setProperty("ob_table_min_rslist_refresh_interval_millis", "1");
+
+        final ObTableClient obTableClient = ObTableClientTestUtil.newTestClient();
+        obTableClient.setMetadataRefreshInterval(100);
+        obTableClient.addProperty(Property.RPC_CONNECT_TIMEOUT.getKey(), "800");
+        obTableClient.addProperty(Property.RPC_LOGIN_TIMEOUT.getKey(), "800");
+        obTableClient.addProperty(Property.SERVER_CONNECTION_POOL_SIZE.getKey(), "1");
+        obTableClient.addProperty(Property.RPC_EXECUTE_TIMEOUT.getKey(), "3000");
+        obTableClient.addProperty(Property.RUNTIME_BATCH_MAX_WAIT.getKey(), "3000");
+        obTableClient.addProperty(Property.RUNTIME_BATCH_EXECUTOR.getKey(), "32");
+        obTableClient.addProperty(Property.RPC_OPERATION_TIMEOUT.getKey(), "3000");
+        obTableClient.init();
+        obTableClient.addRowKeyElement("test_auto_increment_rowkey", new String[] { "c1", "c2" });
+        obTableClient.addRowKeyElement("test_auto_increment_not_rowkey", new String[] { "c1" });
+
+        this.client = obTableClient;
+        syncRefreshMetaHelper(obTableClient);
+    }
+
+    private long getMaxAccessTime(ObTableClient client) throws Exception {
+        Class clientClass = client.getClass();
+        Field field = clientClass.getDeclaredField("serverRoster");
+        field.setAccessible(true);
+        ServerRoster serverRoster = (ServerRoster) field.get(client);
+        long resTime = 0;
+        for (ObServerAddr addr : serverRoster.getMembers()) {
+            resTime = Math.max(resTime, addr.getLastAccessTime());
+        }
+        return resTime;
+    }
+
+    @BeforeClass
+    public static void testVersion() throws Exception {
+        final ObTableClient obTableClient = ObTableClientTestUtil.newTestClient();
+        obTableClient.init();
+        if (ObGlobal.OB_VERSION <= 0) {
+            // ob version is invalid
+            Assert.assertTrue(false);
+        } else if (ObGlobal.OB_VERSION != 4) {
+            // todo: only support in 4.x currently
+            Assert.assertTrue(false);
+        }
+    }
+
+    @Test
+    // Test auto increment on rowkey
+    public void testAutoIncrementRowkey() throws Exception {
+        // use sql to drop table
+        Connection connection = ObTableClientTestUtil.getConnection();
+        Statement statement = connection.createStatement();
+
+        try {
+            client.insert("test_auto_increment_rowkey", new Object[] { 0, 1 }, new String[] {"c3"},
+                    new Object[] { 1 });
+
+            TableQuery tableQuery = client.query("test_auto_increment_rowkey");
+            tableQuery.select("c1", "c2", "c3");
+            ObTableValueFilter filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 1);
+            tableQuery.setFilter(filter);
+            QueryResultSet result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            Map<String, Object> value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(1, value.get("c3"));
+
+            // test insert use user value
+            client.insert("test_auto_increment_rowkey", new Object[] { 100, 1 }, new String[] {"c3"},
+                    new Object[] { 1 });
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 100);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(1, value.get("c3"));
+
+            // test insert sync global auto inc val
+            client.insert("test_auto_increment_rowkey", new Object[] { 0, 1 }, new String[] {"c3"},
+                    new Object[] { 1 });
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 101);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(1, value.get("c3"));
+
+            // test delete
+            client.delete("test_auto_increment_rowkey", new Object[] { 101, 1 });
+
+            // test confirm delete
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 101);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertFalse(result.next());
+
+            // test update
+            ObTableValueFilter filter_3 = compareVal(ObCompareOp.EQ, "c3", 1);
+
+            MutationResult updateResult = client.update("test_auto_increment_rowkey")
+                    .setRowKey(colVal("c1", 1), colVal("c2", 1)).setFilter(filter_3)
+                    .addMutateRow(row(colVal("c3", 5))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 1);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(5, value.get("c3"));
+
+            // test replace not exist, insert
+            MutationResult theResult = client.replace("test_auto_increment_rowkey")
+                    .setRowKey(colVal("c1", 0), colVal("c2",1))
+                    .addMutateRow(row(colVal("c3", 2))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 102);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(2, value.get("c3"));
+
+            // test replace exist, replace
+            theResult  = client.replace("test_auto_increment_rowkey")
+                    .setRowKey(colVal("c1", 101), colVal("c2",1))
+                    .addMutateRow(row(colVal("c3", 20))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 101);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(20, value.get("c3"));
+
+            // test insertup not exist, insert
+            theResult = client.insertOrUpdate("test_auto_increment_rowkey")
+                    .setRowKey(colVal("c1", 0), colVal("c2", 1))
+                    .addMutateRow(row(colVal("c3", 5))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 103);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(5, value.get("c3"));
+
+            // test insertup exist, update
+            theResult = client.insertOrUpdate("test_auto_increment_rowkey")
+                    .setRowKey(colVal("c1", 103), colVal("c2", 1))
+                    .addMutateRow(row(colVal("c3", 50))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 103);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(50, value.get("c3"));
+
+            // test insertup exist, update again
+            theResult = client.insertOrUpdate("test_auto_increment_rowkey")
+                    .setRowKey(colVal("c1", 103), colVal("c2", 1))
+                    .addMutateRow(row(colVal("c3", 50))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 103);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(50, value.get("c3"));
+
+            // test increment not exist, insert
+            value = client.increment("test_auto_increment_rowkey", new Object[] {0, 1},
+                    new String[] { "c3" }, new Object[] { 6 }, true);
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 104);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(6, value.get("c3"));
+
+            // test increment exist, increment
+            value = client.increment("test_auto_increment_rowkey", new Object[] {104, 1},
+                    new String[] { "c3" }, new Object[] { 6 }, true);
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 104);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(12, value.get("c3"));
+
+            // test illegal increment on auto increment column
+            try {
+                value = client.increment("test_auto_increment_rowkey", new Object[] {104, 1},
+                        new String[] { "c1" }, new Object[] { 1 }, true);
+            } catch (ObTableException e) {
+                assertNotNull(e);
+                assertEquals(ResultCodes.OB_NOT_SUPPORTED.errorCode, e.getErrorCode());
+            }
+
+            // test append not exist, insert
+            Map<String, Object> res = client.append("test_auto_increment_rowkey", new Object[] { 0, 1 },
+                    new String[] { "c4" },  new Object[] { "a" } , true);
+
+            tableQuery = client.query("test_auto_increment_rowkey");
+            tableQuery.select("c1", "c2", "c3", "c4");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 105);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals("a", value.get("c4"));
+
+            // test append exist, append
+            res = client.append("test_auto_increment_rowkey", new Object[] { 105, 1 },
+                    new String[] { "c4" },  new Object[] { "b" } , true);
+
+            tableQuery = client.query("test_auto_increment_rowkey");
+            tableQuery.select("c1", "c2", "c3", "c4");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 105);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals("ab", value.get("c4"));
+        } finally { // use sql to drop table
+            statement.execute("drop table test_auto_increment_rowkey");
+        }
+    }
+
+    @Test
+    // Test auto increment on not rowkey
+    public void testAutoIncrementNotRowkey() throws Exception {
+        // use sql to drop table
+        Connection connection = ObTableClientTestUtil.getConnection();
+        Statement statement = connection.createStatement();
+
+        try {
+            client.insert("test_auto_increment_not_rowkey", new Object[] { 1 }, new String[] { "c2" },
+                    new Object[] { 1 });
+
+            TableQuery tableQuery = client.query("test_auto_increment_not_rowkey");
+            tableQuery.select("c1", "c2", "c3");
+            ObTableValueFilter filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 1);
+            tableQuery.setFilter(filter);
+            QueryResultSet result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            Map<String, Object> value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(1, value.get("c3"));
+
+            // test insert use user value
+            client.insert("test_auto_increment_not_rowkey", new Object[] { 2 }, new String[] { "c2", "c3"},
+                    new Object[] { 1, 100 });
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 2);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(100, value.get("c3"));
+
+            // test insert sync global auto inc val
+            client.insert("test_auto_increment_not_rowkey", new Object[] { 3 }, new String[] {"c2"},
+                    new Object[] { 1 });
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 3);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(101, value.get("c3"));
+
+            // test delete
+            client.delete("test_auto_increment_not_rowkey", new Object[] { 1 });
+
+            // test confirm delete
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 1);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertFalse(result.next());
+
+            // test update
+            ObTableValueFilter filter_3 = compareVal(ObCompareOp.EQ, "c2", 1);
+
+            MutationResult updateResult = client.update("test_auto_increment_not_rowkey")
+                    .setRowKey(colVal("c1", 3)).setFilter(filter_3)
+                    .addMutateRow(row(colVal("c3", 5))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 3);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(5, value.get("c3"));
+
+            // test replace not exist, insert
+            MutationResult theResult = client.replace("test_auto_increment_not_rowkey")
+                    .setRowKey(colVal("c1", 4))
+                    .addMutateRow(row(colVal("c2", 1))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 4);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(102, value.get("c3"));
+
+            // test replace exist, replace
+            theResult  = client.replace("test_auto_increment_not_rowkey")
+                    .setRowKey(colVal("c1", 3))
+                    .addMutateRow(row(colVal("c3", 20), colVal("c2", 1))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 3);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(20, value.get("c3"));
+
+            // test insertup not exist, insert
+            theResult = client.insertOrUpdate("test_auto_increment_not_rowkey")
+                    .setRowKey(colVal("c1", 5))
+                    .addMutateRow(row(colVal("c2", 1))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 5);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(103, value.get("c3"));
+
+            // test insertup exist, update
+            theResult = client.insertOrUpdate("test_auto_increment_not_rowkey")
+                    .setRowKey(colVal("c1", 5))
+                    .addMutateRow(row(colVal("c3", 50), colVal("c2", 1))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 5);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(50, value.get("c3"));
+
+            // test insertup exist, update again
+            theResult = client.insertOrUpdate("test_auto_increment_not_rowkey")
+                    .setRowKey(colVal("c1", 5))
+                    .addMutateRow(row(colVal("c3", 50), colVal("c2", 1))).execute();
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 5);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(1, value.get("c2"));
+            Assert.assertEquals(50, value.get("c3"));
+
+            // test increment not exist, insert
+            value = client.increment("test_auto_increment_not_rowkey", new Object[] {6},
+                    new String[] { "c2" }, new Object[] { 6 }, true);
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 6);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(6, value.get("c2"));
+            Assert.assertEquals(104, value.get("c3"));
+
+            // test increment exist, increment
+            value = client.increment("test_auto_increment_not_rowkey", new Object[] {6},
+                    new String[] { "c2" }, new Object[] { 6 }, true);
+
+            tableQuery.select("c1", "c2", "c3");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 6);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(12, value.get("c2"));
+            Assert.assertEquals(104, value.get("c3"));
+
+            // test illegal increment on auto increment column
+            try {
+                value = client.increment("test_auto_increment_not_rowkey", new Object[] {6},
+                        new String[] { "c3" }, new Object[] { 1 }, true);
+            } catch (ObTableException e) {
+                assertNotNull(e);
+                assertEquals(ResultCodes.OB_NOT_SUPPORTED.errorCode, e.getErrorCode());
+            }
+
+            // test append not exist, insert
+            Map<String, Object> res = client.append("test_auto_increment_not_rowkey", new Object[] { 200 },
+                    new String[] { "c4" },  new Object[] { "a" } , true);
+
+            tableQuery = client.query("test_auto_increment_not_rowkey");
+            tableQuery.select("c1", "c2", "c3", "c4");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 200);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(106, value.get("c3"));
+            Assert.assertEquals("a", value.get("c4"));
+
+            // test append exist, append
+            res = client.append("test_auto_increment_not_rowkey", new Object[] { 200 },
+                    new String[] { "c4" },  new Object[] { "b" } , true);
+
+            tableQuery = client.query("test_auto_increment_not_rowkey");
+            tableQuery.select("c1", "c2", "c3", "c4");
+            filter = new ObTableValueFilter(ObCompareOp.EQ, "c1", 200);
+            tableQuery.setFilter(filter);
+            result = tableQuery.execute();
+            Assert.assertTrue(result.next());
+            value = result.getRow();
+            Assert.assertEquals(106, value.get("c3"));
+            Assert.assertEquals("ab", value.get("c4"));
+        } finally { // use sql to drop table
+            statement.execute("drop table test_auto_increment_not_rowkey");
+        }
+    }
+}
